+++ front-end/app/src/main/java/com/example/grocerymanager/Item.java
@@ -5,9 +5,9 @@ public class Item {
     private String expiry;
     private int quantity;
     private int itemId;
-    private  int UPC;
+    private  long UPC;
 
-    public Item(String name, String expiry, int quantity, int itemId, int UPC) {
+    public Item(String name, String expiry, int quantity, int itemId, long UPC) {
         this.name = name;
         this.expiry = expiry;
         this.quantity = quantity;
@@ -42,8 +42,8 @@ public class Item {
     public void setItemId(int itemId){
         this.itemId = itemId;
     }
-    public int getUPC(){return UPC;}
-    public void setUPC(int UPC){
+    public long getUPC(){return UPC;}
+    public void setUPC(long UPC){
         this.UPC = UPC;
     }
 }
